# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SVC Ecommerce API",
        "description": "API documentation for SVC Coffee Ecommerce Platform - Complete e-commerce solution with cart, checkout, and order management",
        "contact": {
            "name": "SVC Development Team",
            "email": "usman@workforcecommerce.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://workforcecommerce.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://staging.workforcecommerce.com",
            "description": "Staging Server"
        },
        {
            "url": "https://api.workforcecommerce.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/checkout/create-order": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Create order from cart",
                "description": "Create a new order from cart. Supports both guest and registered customer checkout.",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cart_id",
                                    "shipping_first_name",
                                    "shipping_last_name",
                                    "shipping_address_line1",
                                    "shipping_city",
                                    "shipping_postal_code",
                                    "shipping_country",
                                    "shipping_phone",
                                    "shipping_method",
                                    "payment_method"
                                ],
                                "properties": {
                                    "cart_id": {
                                        "type": "string",
                                        "example": "abc123xyz"
                                    },
                                    "customer_id": {
                                        "description": "Customer UUID (null for guest checkout)",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "guest_email": {
                                        "description": "Required if customer_id is null",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "guest_name": {
                                        "description": "Required if customer_id is null",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "guest_phone": {
                                        "type": "string",
                                        "example": "+1234567890",
                                        "nullable": true
                                    },
                                    "shipping_first_name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "John"
                                    },
                                    "shipping_last_name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Doe"
                                    },
                                    "shipping_address_line1": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "123 Main Street"
                                    },
                                    "shipping_address_line2": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Apt 4B",
                                        "nullable": true
                                    },
                                    "shipping_city": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "New York"
                                    },
                                    "shipping_state": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "NY",
                                        "nullable": true
                                    },
                                    "shipping_postal_code": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "10001"
                                    },
                                    "shipping_country": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "United States"
                                    },
                                    "shipping_phone": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "+1234567890"
                                    },
                                    "billing_same_as_shipping": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "billing_first_name": {
                                        "description": "Required if billing_same_as_shipping is false",
                                        "type": "string",
                                        "example": "John",
                                        "nullable": true
                                    },
                                    "billing_last_name": {
                                        "type": "string",
                                        "example": "Doe",
                                        "nullable": true
                                    },
                                    "billing_address_line1": {
                                        "type": "string",
                                        "example": "456 Business Ave",
                                        "nullable": true
                                    },
                                    "billing_address_line2": {
                                        "type": "string",
                                        "example": "Suite 100",
                                        "nullable": true
                                    },
                                    "billing_city": {
                                        "type": "string",
                                        "example": "Los Angeles",
                                        "nullable": true
                                    },
                                    "billing_state": {
                                        "type": "string",
                                        "example": "CA",
                                        "nullable": true
                                    },
                                    "billing_postal_code": {
                                        "type": "string",
                                        "example": "90001",
                                        "nullable": true
                                    },
                                    "billing_country": {
                                        "type": "string",
                                        "example": "United States",
                                        "nullable": true
                                    },
                                    "shipping_method": {
                                        "type": "string",
                                        "enum": [
                                            "standard",
                                            "express",
                                            "overnight",
                                            "free"
                                        ],
                                        "example": "standard"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "credit_card",
                                            "paypal",
                                            "stripe",
                                            "cash_on_delivery"
                                        ],
                                        "example": "credit_card"
                                    },
                                    "customer_notes": {
                                        "type": "string",
                                        "example": "Please ring doorbell",
                                        "nullable": true
                                    },
                                    "coupon_code": {
                                        "type": "string",
                                        "example": "SUMMER2025",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "order_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "order_number": {
                                                    "type": "string",
                                                    "example": "ORD-2025-00001"
                                                },
                                                "total_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 659.97
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "USD"
                                                },
                                                "payment_required": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cart is empty or insufficient stock",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cart is empty"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Failed to create order"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Enter your API key",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "apiKey": []
        }
    ]
}